pyproject.toml
[tool.poetry]
name = "ijburg_futsal"
version = "0.1.0"
description = ""
authors = ["Daniel Bradburn <daniel@crunchrapps.com>"]

[tool.poetry.dependencies]
python = "^3.8"
funcoperators = "^1.1.3"
toolz = "^0.11.1"
pandas = "^1.1.2"
PyYAML = "^6.0.1"
munch = "^4.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
